requirements.txt
asgiref==3.7.2
black==23.9.1
click==8.1.7
Django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
flake8==6.1.0
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.3.0
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.2
pytz==2023.3.post1
sqlparse==0.4.4
